pyproject.toml
[tool.poetry]
name = "shortcut-changelog-generator"
version = "1.0.0"
description = ""
authors = ["Diogenes Castro"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
GitPython = "^3.1.24"
requests = "^2.27.1"
python-dotenv = "^0.19.2"
termcolor = "^1.1.0"
python-telegram-bot = "^13.10"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
